refactor(args): share fuzzyMatch scoreName helper
Extract scoreName into utils/fuzzyMatch for reuse; extend segment split
to include whitespace for natural-language install labels.
